move groupby from rules to utilities/iterables
